// src/components/ui/material/material-elements.d.ts
export declare const MaterialElements: {
    FilledButton: any;
    OutlinedButton: any;
    TextButton: any;
    ElevatedButton: any;
    FilledTonalButton: any;
    Icon: any;
    IconButton: any;
    FilledTextField: any;
    OutlinedTextField: any;
    Dialog: any;
    FilledSelect: any;
    OutlinedSelect: any;
    SelectOption: any;
    Switch: any;
    Tabs: any;
    PrimaryTab: any;
    SecondaryTab: any;
    Fab: any;
    BrandedFab: any;
    ChipSet: any;
    AssistChip: any;
    FilterChip: any;
    InputChip: any;
    SuggestionChip: any;
    List: any;
    ListItem: any;
    Ripple: any;
};